+++ src/ak_utils.c
@@ -95,5 +95,5 @@ ak_parse_date(const char * __restrict input,
 AK_EXPORT
 int
 ak_digitsize(size_t number) {
-  return floor(log10(number)) + 1;
+  return (int)floor(log10(number)) + 1;
 }
